chore(agents): remove non-mcp version
