--- dimos/agents/fixtures/test_image.json
@@ -1,23 +1,0 @@
-{
-  "responses": [
-    {
-      "content": "",
-      "tool_calls": [
-        {
-          "name": "take_a_picture",
-          "args": {},
-          "id": "call_11vyrbtZXlsEgoKuQs8jRjP8",
-          "type": "tool_call"
-        }
-      ]
-    },
-    {
-      "content": "I have taken a picture, let me analyze it to provide a description.\nBased on the image captured, the setting resembles a \"cafe\". It shows an indoor space with tables, chairs, and a warm, inviting atmosphere typical of a cafe environment, where people might gather for drinks and conversation.",
-      "tool_calls": []
-    },
-    {
-      "content": "The image depicts a \"cafe\" setting. It shows people sitting outside at tables, likely enjoying drinks or meals, with a cozy and inviting ambiance typical of a cafe.",
-      "tool_calls": []
-    }
-  ]
-}
