+++ dimos/agents/skills/demo_google_maps_skill.py
@@ -13,7 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from dimos.agents.agent import Agent
+from dimos.agents.mcp.mcp_client import McpClient
+from dimos.agents.mcp.mcp_server import McpServer
 from dimos.agents.skills.demo_robot import DemoRobot
 from dimos.agents.skills.google_maps_skill_container import GoogleMapsSkillContainer
 from dimos.core.blueprints import autoconnect
@@ -21,5 +22,6 @@ from dimos.core.blueprints import autoconnect
 demo_google_maps_skill = autoconnect(
     DemoRobot.blueprint(),
     GoogleMapsSkillContainer.blueprint(),
-    Agent.blueprint(),
+    McpServer.blueprint(),
+    McpClient.blueprint(),
 )
