+++ dimos/agents/skills/demo_skill.py
@@ -13,11 +13,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from dimos.agents.agent import Agent
+from dimos.agents.mcp.mcp_client import McpClient
+from dimos.agents.mcp.mcp_server import McpServer
 from dimos.agents.skills.demo_calculator_skill import DemoCalculatorSkill
 from dimos.core.blueprints import autoconnect
 
 demo_skill = autoconnect(
     DemoCalculatorSkill.blueprint(),
-    Agent.blueprint(),
+    McpServer.blueprint(),
+    McpClient.blueprint(),
 )
