+++ dimos/e2e_tests/test_dimos_cli_e2e.py
@@ -20,13 +20,13 @@ import pytest
 @pytest.mark.slow
 def test_dimos_skills(lcm_spy, start_blueprint, human_input) -> None:
     lcm_spy.save_topic("/agent")
-    lcm_spy.save_topic("/rpc/Agent/on_system_modules/res")
+    lcm_spy.save_topic("/rpc/McpClient/on_system_modules/res")
     lcm_spy.save_topic("/rpc/DemoCalculatorSkill/sum_numbers/req")
     lcm_spy.save_topic("/rpc/DemoCalculatorSkill/sum_numbers/res")
 
     start_blueprint("run", "demo-skill")
 
-    lcm_spy.wait_for_saved_topic("/rpc/Agent/on_system_modules/res")
+    lcm_spy.wait_for_saved_topic("/rpc/McpClient/on_system_modules/res")
 
     human_input("what is 52983 + 587237")
