+++ dimos/e2e_tests/test_scan_and_follow_person.py
@@ -40,8 +40,8 @@ def test_scan_and_person_follow(
         "unitree-go2-agentic",
     )
 
-    lcm_spy.save_topic("/rpc/Agent/on_system_modules/res")
-    lcm_spy.wait_for_saved_topic("/rpc/Agent/on_system_modules/res", timeout=120.0)
+    lcm_spy.save_topic("/rpc/McpClient/on_system_modules/res")
+    lcm_spy.wait_for_saved_topic("/rpc/McpClient/on_system_modules/res", timeout=120.0)
 
     time.sleep(5)
