+++ dimos/perception/demo_object_scene_registration.py
@@ -13,7 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from dimos.agents.agent import Agent
+from dimos.agents.mcp.mcp_client import McpClient
+from dimos.agents.mcp.mcp_server import McpServer
 from dimos.core.blueprints import autoconnect
 from dimos.hardware.sensors.camera.realsense.camera import RealSenseCamera
 from dimos.hardware.sensors.camera.zed.compat import ZEDCamera
@@ -34,5 +35,6 @@ demo_object_scene_registration = autoconnect(
     camera_module,
     ObjectSceneRegistrationModule.blueprint(target_frame="world", prompt_mode=YoloePromptMode.LRPC),
     FoxgloveBridge.blueprint(),
-    Agent.blueprint(),
+    McpServer.blueprint(),
+    McpClient.blueprint(),
 ).global_config(viewer="foxglove")
