+++ dimos/robot/drone/blueprints/agentic/drone_agentic.py
@@ -20,7 +20,8 @@ Composes on top of drone_basic (connection + camera + vis) and adds
 tracking, mapping skills, and an LLM agent.
 """
 
-from dimos.agents.agent import Agent
+from dimos.agents.mcp.mcp_client import McpClient
+from dimos.agents.mcp.mcp_server import McpServer
 from dimos.agents.skills.google_maps_skill_container import GoogleMapsSkillContainer
 from dimos.agents.skills.osm import OsmSkill
 from dimos.agents.web_human_input import WebInput
@@ -44,7 +45,8 @@ drone_agentic = autoconnect(
     DroneTrackingModule.blueprint(outdoor=False),
     GoogleMapsSkillContainer.blueprint(),
     OsmSkill.blueprint(),
-    Agent.blueprint(system_prompt=DRONE_SYSTEM_PROMPT, model="gpt-4o"),
+    McpServer.blueprint(),
+    McpClient.blueprint(system_prompt=DRONE_SYSTEM_PROMPT, model="gpt-4o"),
     WebInput.blueprint(),
 ).remappings(
     [
