+++ dimos/robot/unitree/g1/blueprints/agentic/_agentic_skills.py
@@ -15,7 +15,8 @@
 
 """Agentic skills used by higher-level G1 blueprints."""
 
-from dimos.agents.agent import Agent
+from dimos.agents.mcp.mcp_client import McpClient
+from dimos.agents.mcp.mcp_server import McpServer
 from dimos.agents.skills.navigation import NavigationSkillContainer
 from dimos.agents.skills.speak_skill import SpeakSkill
 from dimos.core.blueprints import autoconnect
@@ -23,7 +24,8 @@ from dimos.robot.unitree.g1.skill_container import UnitreeG1SkillContainer
 from dimos.robot.unitree.g1.system_prompt import G1_SYSTEM_PROMPT
 
 _agentic_skills = autoconnect(
-    Agent.blueprint(system_prompt=G1_SYSTEM_PROMPT),
+    McpServer.blueprint(),
+    McpClient.blueprint(system_prompt=G1_SYSTEM_PROMPT),
     NavigationSkillContainer.blueprint(),
     SpeakSkill.blueprint(),
     UnitreeG1SkillContainer.blueprint(),
