+++ dimos/robot/unitree/go2/blueprints/agentic/unitree_go2_agentic.py
@@ -13,14 +13,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from dimos.agents.agent import Agent
+from dimos.agents.mcp.mcp_client import McpClient
+from dimos.agents.mcp.mcp_server import McpServer
 from dimos.core.blueprints import autoconnect
 from dimos.robot.unitree.go2.blueprints.agentic._common_agentic import _common_agentic
 from dimos.robot.unitree.go2.blueprints.smart.unitree_go2_spatial import unitree_go2_spatial
 
 unitree_go2_agentic = autoconnect(
     unitree_go2_spatial,
-    Agent.blueprint(),
+    McpServer.blueprint(),
+    McpClient.blueprint(),
     _common_agentic,
 )
